Added project guess a number
Simple console game where you have to guess the number. I has 3 levels of difficulty starting from easy then medium and hard each with lower chances to guess the number.
The game also has two levels as the first one is from numbers 1-100 and if you successfully complete it you are going to the second one from numbers 1-200.
After successful completion of the second level you will be prompted to YouTube to listen the "winners song".